Move PcesConfig to event-intake module
