+++ hedera-state-validator/src/main/java/module-info.java
@@ -38,6 +38,7 @@ module com.hedera.state.validator {
     requires org.hiero.base.crypto;
     requires org.hiero.base.utility;
     requires org.hiero.consensus.concurrent;
+    requires org.hiero.consensus.event.intake;
     requires org.hiero.consensus.metrics;
     requires org.hiero.consensus.model;
     requires org.hiero.consensus.utility;
