+++ platform-sdk/consensus-event-intake/src/main/java/org/hiero/consensus/event/intake/config/PcesConfig.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.platform.event.preconsensus;
+package org.hiero.consensus.event.intake.config;
 
 import com.swirlds.config.api.ConfigData;
 import com.swirlds.config.api.ConfigProperty;
