+++ platform-sdk/consensus-event-intake/src/main/java/org/hiero/consensus/event/intake/config/PcesFileWriterType.java
@@ -1,0 +1,11 @@
+// SPDX-License-Identifier: Apache-2.0
+package org.hiero.consensus.event.intake.config;
+
+/**
+ * Represents a type of writers for PCES files.
+ */
+public enum PcesFileWriterType {
+    OUTPUT_STREAM,
+    FILE_CHANNEL,
+    FILE_CHANNEL_SYNC;
+}
