+++ platform-sdk/swirlds-common/src/testFixtures/java/module-info.java
@@ -26,9 +26,9 @@ open module com.swirlds.common.test.fixtures {
     requires transitive org.hiero.consensus.model;
     requires com.swirlds.config.extensions.test.fixtures;
     requires com.swirlds.logging;
-    requires com.swirlds.platform.core;
     requires com.swirlds.virtualmap;
     requires org.hiero.base.utility.test.fixtures;
+    requires org.hiero.consensus.event.intake;
     requires org.hiero.consensus.utility;
     requires org.apache.logging.log4j.core;
     requires org.apache.logging.log4j;
