+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/cli/CrystalTransplantCommand.java
@@ -26,7 +26,6 @@ import com.swirlds.config.api.Configuration;
 import com.swirlds.config.api.ConfigurationBuilder;
 import com.swirlds.config.extensions.sources.LegacyFileConfigSource;
 import com.swirlds.platform.cli.utils.HederaUtils;
-import com.swirlds.platform.event.preconsensus.PcesConfig;
 import com.swirlds.platform.state.SavedStateUtils;
 import com.swirlds.platform.state.snapshot.SavedStateInfo;
 import com.swirlds.platform.state.snapshot.SavedStateMetadata;
@@ -47,6 +46,7 @@ import java.util.Scanner;
 import java.util.stream.Collectors;
 import org.hiero.base.constructable.ConstructableRegistryException;
 import org.hiero.base.crypto.Hash;
+import org.hiero.consensus.event.intake.config.PcesConfig;
 import org.hiero.consensus.metrics.noop.NoOpMetrics;
 import org.hiero.consensus.model.node.NodeId;
 import org.hiero.consensus.roster.RosterDiff;
