+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/event/preconsensus/BestEffortPcesFileCopy.java
@@ -17,6 +17,7 @@ import java.util.List;
 import java.util.stream.Stream;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.hiero.consensus.event.intake.config.PcesConfig;
 import org.hiero.consensus.model.node.NodeId;
 
 /**
