+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/event/preconsensus/PcesFile.java
@@ -17,6 +17,7 @@ import java.time.format.DateTimeParseException;
 import java.util.Objects;
 import java.util.stream.Stream;
 import org.hiero.base.utility.NonCryptographicHashing;
+import org.hiero.consensus.event.intake.config.PcesFileWriterType;
 
 /**
  * <p>
