+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/event/preconsensus/PcesFileWriterFactory.java
@@ -6,24 +6,26 @@ import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
 import java.util.List;
+import org.hiero.consensus.event.intake.config.PcesFileWriterType;
 
 /**
- * Represents a type of writers for PCES files.
+ * Factory for creating PcesFileWriter instances
  */
-public enum PcesFileWriterType {
-    OUTPUT_STREAM,
-    FILE_CHANNEL,
-    FILE_CHANNEL_SYNC;
+public class PcesFileWriterFactory {
+
+    private PcesFileWriterFactory() {}
 
     /**
      * Creates the right instance of the PcesFileWriter for the type represented by this enum
      *
+     * @param writerType the type of writer to create
      * @param path the path to the file to write to
      * @return the writer for writing PCES files
      * @throws IOException in case of error when creating the writer
      */
-    public PcesFileWriter createWriter(@NonNull final Path path) throws IOException {
-        return switch (this) {
+    public static PcesFileWriter createWriter(@NonNull final PcesFileWriterType writerType, @NonNull final Path path)
+            throws IOException {
+        return switch (writerType) {
             case OUTPUT_STREAM -> new PcesOutputStreamFileWriter(path);
             case FILE_CHANNEL -> new PcesFileChannelWriter(path);
             case FILE_CHANNEL_SYNC -> new PcesFileChannelWriter(path, List.of(StandardOpenOption.DSYNC));
