+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/event/EventMigrationTest.java
@@ -4,7 +4,6 @@ package com.swirlds.platform.event;
 import com.swirlds.common.io.IOIterator;
 import com.swirlds.common.io.utility.NoOpRecycleBin;
 import com.swirlds.config.extensions.test.fixtures.TestConfigBuilder;
-import com.swirlds.platform.event.preconsensus.PcesConfig_;
 import com.swirlds.platform.event.preconsensus.PcesFileReader;
 import com.swirlds.platform.event.preconsensus.PcesFileTracker;
 import com.swirlds.platform.recovery.internal.EventStreamSingleFileIterator;
@@ -24,6 +23,7 @@ import org.hiero.base.constructable.ConstructableRegistry;
 import org.hiero.base.constructable.ConstructableRegistryException;
 import org.hiero.base.crypto.Hash;
 import org.hiero.consensus.crypto.DefaultEventHasher;
+import org.hiero.consensus.event.intake.config.PcesConfig_;
 import org.hiero.consensus.model.event.EventDescriptorWrapper;
 import org.hiero.consensus.model.event.PlatformEvent;
 import org.junit.jupiter.api.Assertions;
